feat(config): load global config excluding CLI subcommand overrides
- Extract global configuration loading logic to `config_loader` module.
- Filter out empty CLI flatten group layers that could overwrite file/env values.
- Preserve generated discovery flags like `--config-path` when loading global args.
- Adapt main.rs to use new config loading function, improving config discovery.
- Add tests for loading global args without CLI overrides and honoring config path.

This change improves configuration handling by isolating CLI subcommands from global config loading, preventing unintended overrides and enabling better config file discovery.
