+++ docs/github-token.md
@@ -5,11 +5,11 @@ Follow these steps to create one:
 
 1. Visit <https://github.com/settings/tokens> and choose **Generate new token**.
    GitHub may prompt for a classic or fine‑grained token – either works.
-2. Give the token a note and set an expiration.
-3. Under **Select scopes**, enable `public_repo`. If private repositories are
+1. Give the token a note and set an expiration.
+1. Under **Select scopes**, enable `public_repo`. If private repositories are
    required, select the broader `repo` scope instead.
-4. Click **Generate token** and copy the value.
-5. Export the token as `VK_GITHUB_TOKEN` (or `GITHUB_TOKEN`):
+1. Click **Generate token** and copy the value.
+1. Export the token as `VK_GITHUB_TOKEN` (or `GITHUB_TOKEN`):
 
 ```bash
 export VK_GITHUB_TOKEN=YOUR_TOKEN
