+++ docs/ortho-config-users-guide.md
@@ -583,7 +583,7 @@ following steps:
 1. Builds a `figment` configuration profile. A defaults provider constructed
    from the `#[ortho_config(default = …)]` attributes is added first.
 
-2. Attempts to load a configuration file. Candidate file paths are searched in
+1. Attempts to load a configuration file. Candidate file paths are searched in
    the following order:
 
    1. If provided, a path supplied via the CLI flag generated by the
@@ -592,30 +592,30 @@ following steps:
       `APP_CONFIG_PATH` or `CONFIG_PATH`) takes precedence; see
       [Config path override](#config-path-override).
 
-   2. A dotfile named `.<prefix>.toml` in the current working directory.
+   1. A dotfile named `.<prefix>.toml` in the current working directory.
 
-   3. A dotfile of the same name in the user's home directory.
+   1. A dotfile of the same name in the user's home directory.
 
-   4. On Unix‑like systems, the XDG configuration directory (e.g.
+   1. On Unix‑like systems, the XDG configuration directory (e.g.
       `~/.config/app/config.toml`) is searched using the `xdg` crate; on
       Windows, the `%APPDATA%` and `%LOCALAPPDATA%` directories are checked.
 
-   5. If the `json5` or `yaml` features are enabled, files with `.json`,
+   1. If the `json5` or `yaml` features are enabled, files with `.json`,
       `.json5`, `.yaml`, or `.yml` extensions are also considered in these
       locations.
 
-3. Adds an environment provider using the prefix specified on the struct. Keys
+1. Adds an environment provider using the prefix specified on the struct. Keys
    are upper‑cased and nested fields use double underscores (`__`) to separate
    components.
 
-4. Adds a provider containing the CLI values (captured as `Option<T>` fields)
+1. Adds a provider containing the CLI values (captured as `Option<T>` fields)
    as the final layer.
 
-5. Merges collection fields according to the configured `merge_strategy`, using
+1. Merges collection fields according to the configured `merge_strategy`, using
    `append` by default for vectors, `keyed` by default for maps, and `replace`
    when later sources should override earlier ones.
 
-6. Attempts to extract the merged configuration into the concrete struct. On
+1. Attempts to extract the merged configuration into the concrete struct. On
    success it returns the completed configuration; otherwise an `OrthoError` is
    returned.
 
@@ -666,13 +666,13 @@ earlier ones. The precedence, from lowest to highest, is:
 1. **Application‑defined defaults** – values provided via `default` attributes
    or `Option<T>` fields are considered defaults.
 
-2. **Configuration file** – values from a TOML (or JSON5/YAML) file loaded from
+1. **Configuration file** – values from a TOML (or JSON5/YAML) file loaded from
    one of the paths listed above.
 
-3. **Environment variables** – variables prefixed with the struct's `prefix`
+1. **Environment variables** – variables prefixed with the struct's `prefix`
    (e.g. `APP_PORT`, `APP_DATABASE__URL`) override file values.
 
-4. **Command‑line arguments** – values parsed by `clap` override all other
+1. **Command‑line arguments** – values parsed by `clap` override all other
    sources.
 
 Nested structs are flattened in the environment namespace by joining field
@@ -961,9 +961,9 @@ remains available for generated defaults/documentation metadata.
 **Precedence with the attribute (lowest to highest):**
 
 1. Struct default (`#[ortho_config(default = ...)]` or inferred from clap)
-2. Configuration file
-3. Environment variable
-4. Explicit CLI override (e.g. `--punctuation "?"`)
+1. Configuration file
+1. Environment variable
+1. Explicit CLI override (e.g. `--punctuation "?"`)
 
 Without `cli_default_as_absent`, the clap default would always beat the file
 and environment layers. With the attribute, calling `greet` without
@@ -1190,15 +1190,15 @@ cargo orthohelp --format man --out-dir target/man --locale en-US
 The generator produces standard man page sections in the canonical order:
 
 1. **NAME** – binary name and one-line description
-2. **SYNOPSIS** – usage pattern with flags
-3. **DESCRIPTION** – expanded about text
-4. **OPTIONS** – CLI flags with types, defaults, and possible values
-5. **ENVIRONMENT** – environment variables mapped to fields
-6. **FILES** – configuration file paths and discovery locations
-7. **PRECEDENCE** – source priority order (defaults → file → env → CLI)
-8. **EXAMPLES** – usage examples from the IR
-9. **SEE ALSO** – related commands and documentation links
-10. **EXIT STATUS** – standard exit codes
+1. **SYNOPSIS** – usage pattern with flags
+1. **DESCRIPTION** – expanded about text
+1. **OPTIONS** – CLI flags with types, defaults, and possible values
+1. **ENVIRONMENT** – environment variables mapped to fields
+1. **FILES** – configuration file paths and discovery locations
+1. **PRECEDENCE** – source priority order (defaults → file → env → CLI)
+1. **EXAMPLES** – usage examples from the IR
+1. **SEE ALSO** – related commands and documentation links
+1. **EXIT STATUS** – standard exit codes
 
 Additional options:
