+++ docs/ortho-config-v0-7-0-migration-guide.md
@@ -18,9 +18,8 @@
 This guide describes how to upgrade applications from `ortho-config` v0.6.0 to
 v0.7.0. The release is additive: it introduces layer composition helpers,
 post-merge hooks, localized CLI output, and additional error-handling
-utilities, without changing the core configuration model. Use the sections
-below to adopt the new helpers where they match your configuration style and
-testing setup.
+utilities, without changing the core configuration model. The sections below
+map the new helpers to matching configuration styles and test setups.
 
 ## At-a-glance breaking changes
 
@@ -47,11 +46,11 @@ ortho_config_macros = "0.7.0"
 ```
 
 1. Update every `ortho_config` and `ortho_config_macros` dependency to `0.7.0`.
-2. Keep format features (`toml`, `json5`, `yaml`) on `ortho_config` as before.
-3. When default features are disabled, enable `serde_json` explicitly whenever
+1. Keep format features (`toml`, `json5`, `yaml`) on `ortho_config` as before.
+1. When default features are disabled, enable `serde_json` explicitly whenever
    `cli_default_as_absent` or selected-subcommand merging helpers are
    required. The default feature set already enables it.
-4. Expect new transitive dependencies (`fluent-bundle`, `fluent-syntax`,
+1. Expect new transitive dependencies (`fluent-bundle`, `fluent-syntax`,
    `unic-langid`, and `tracing`) to land with v0.7.0, as they power CLI
    localization support.[^deps-0-7]
 
@@ -175,7 +174,7 @@ step. Variants that depend on `cli_default_as_absent` should be annotated with
 
 ## 6. Refresh error handling and tests
 
-- Use `is_display_request` before mapping `clap::Error` into your own error
+- Use `is_display_request` before mapping `clap::Error` into application error
   types, so `--help` and `--version` still exit with code 0.[^display-request]
 - `OrthoJsonMergeExt::into_ortho_merge_json()` attributes JSON parsing
   failures to the merge phase while preserving line and column details.
@@ -185,42 +184,42 @@ step. Variants that depend on `cli_default_as_absent` should be annotated with
   updated to use the `rstest-bdd` macros and compile-time tag filters (for
   example, gating YAML scenarios).[^rstest-bdd]
 
-[^deps-0-7]: v0.7.0 adds Fluent and localization dependencies alongside the
+\[^deps-0-7\]: v0.7.0 adds Fluent and localization dependencies alongside the
 existing feature flags in the runtime crate and macro crate metadata. See
 [`ortho_config/Cargo.toml`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/Cargo.toml)
 and
 [`ortho_config_macros/Cargo.toml`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config_macros/Cargo.toml).
-[^compose-layers]: Derived configuration structs now expose
+\[^compose-layers\]: Derived configuration structs now expose
 `compose_layers()` and `compose_layers_from_iter(...)` and return
 `LayerComposition` for staged merging and error aggregation. See
 [`ortho_config_macros/src/derive/load_impl.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config_macros/src/derive/load_impl.rs)
 and
 [`ortho_config/src/declarative.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/declarative.rs).
-[^post-merge]: `PostMergeHook` and `PostMergeContext` enable post-merge
+\[^post-merge\]: `PostMergeHook` and `PostMergeContext` enable post-merge
 adjustments and are invoked when `#[ortho_config(post_merge_hook)]` is
 present. See
 [`ortho_config/src/post_merge.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/post_merge.rs).
-[^localizer]: `Localizer`, `LocalizationArgs`, and `FluentLocalizer` define
+\[^localizer\]: `Localizer`, `LocalizationArgs`, and `FluentLocalizer` define
 the CLI localization surface and provide a Fluent-backed implementation.
 See
 [`ortho_config/src/localizer/mod.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/localizer/mod.rs).
-[^localizeclap]: `localize_clap_error_with_command` rewrites clap error
+\[^localizeclap\]: `localize_clap_error_with_command` rewrites clap error
 messages while keeping the existing rendered tail intact. See
 [`ortho_config/src/localizer/clap_error.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/localizer/clap_error.rs).
-[^cli-defaults]: `cli_default_as_absent` is implemented via the
+\[^cli-defaults\]: `cli_default_as_absent` is implemented via the
 `CliValueExtractor` trait to exclude clap defaults from the CLI layer
 unless explicitly provided. See
 [`ortho_config/src/merge.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/merge.rs).
-[^selectedsubcommand]: `SelectedSubcommandMerge` and
+\[^selectedsubcommand\]: `SelectedSubcommandMerge` and
 `load_globals_and_merge_selected_subcommand` let the chosen subcommand be
 merged without manual `match` scaffolding. See
 [`ortho_config/src/subcommand/selected.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/subcommand/selected.rs).
-[^display-request]: `is_display_request` preserves clap's help and version
+\[^display-request\]: `is_display_request` preserves clap's help and version
 exit behaviour when callers use `try_parse()`. See
 [`ortho_config/src/error.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/error.rs).
-[^json-merge]: `OrthoJsonMergeExt` maps `serde_json::Error` into merge
+\[^json-merge\]: `OrthoJsonMergeExt` maps `serde_json::Error` into merge
 failures while preserving detailed diagnostics. See
 [`ortho_config/src/result_ext.rs`](https://github.com/leynos/ortho-config/blob/v0.7.0/ortho_config/src/result_ext.rs).
-[^rstest-bdd]: The behavioural suites moved from `cucumber-rs` to
+\[^rstest-bdd\]: The behavioural suites moved from `cucumber-rs` to
 `rstest-bdd`, including tag-aware filtering for YAML scenarios. See
 [`docs/adr-002-replace-cucumber-with-rstest-bdd.md`](https://github.com/leynos/ortho-config/blob/v0.7.0/docs/adr-002-replace-cucumber-with-rstest-bdd.md).
